+++ CHANGELOG.md
@@ -1,1 +1,2 @@
+- feat: Add requiresAPI function to allow declaring Google Cloud API dependencies in code.
 - fix(v1): Call onInit for schedule.onRun functions (#1801)